+++ docs/CHANGELOG.rst
@@ -18,7 +18,7 @@ Added
 - Add ExtendedCollectionFilter filter to allow filtering collections by samples
   containing given species, tissue type, outcome or treatment type
 - Add process ``reference-space`` and ``upload-ml-expression``
-- Add a beta version of MACS 2.0 Python process ``macs2-callpeak-beta``
+- Rewrite ``macs2-callpeak`` process to Python
 - Add process ``mutations-table``
 
 Changed